fix: 하드코딩 하지 맙시다

+++ app/src/main/java/com/kinopio/eatgo/presentation/qr/ScanQRActivity.kt
@@ -32,7 +32,7 @@ class ScanQRActivity : AppCompatActivity() {
         else {
             Log.d("review", result.formatName)
 
-            var storeId = result.contents
+            var storeId = result.contents.toInt()
             Log.d("QR", "${storeId}")
             Log.d("QR", "${result.contents}")
             var userId = User.getUserId()
